echoes_of_the_unseen/src/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
google-cloud-firestore>=2.11.0 # Ensure a specific version
google-cloud-aiplatform>=1.25.0 # For Vertex AI (Gemini)
google-cloud-texttospeech # Retaining this as it was present, for standard TTS
google-cloud-speech # Retaining this as it was present
elevenlabs>=0.2.20 # Official SDK
python-dotenv>=0.21.0 # For local environment variable management

# Other useful libraries you might need later
# httpx # For making external HTTP requests if needed directly
